codex/requirements.txt
datasets>=2.14.0
pandas>=2.0.0
tqdm>=4.65.0
pyyaml>=6.0
GitPython>=3.1.0
# Multi-SWE-Bench official evaluation tool
multi-swe-bench>=0.1.0
